Fix For HTML5 Build bug
My first real PR, so I accidentally didn't put this on a new branch. Oops.

Basically, without this, HTML5 seems to only play the audio for the cutscene instead of the webm and audio.

Since commenting this code out breaks desktop builds in return, we will be doing this instead.